pyproject.toml
[tool.poetry]
name = "django-channels"
version = "0.1.0"
description = ""
authors = ["admin1282 <gaikwadvinayak1282@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
django-split-settings = "^1.3.1"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
pytest-django = "^4.8.0"
channels = {extras = ["daphne"], version = "^4.1.0"}
djangorestframework = "^3.15.1"
channels-redis = "^4.2.0"
drf-yasg = "^1.21.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
